backend/nlp-service/requirements.txt
fastapi
uvicorn[standard]
transformers==4.41.2
torch==2.2.2
pydantic
kafka-python
bertopic==0.16.0
